+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add D'Agate 2024 aciclovir ([doi:10.1002/prp2.1193](https://doi.org/10.1002/prp2.1193)) -- term and pre-term neonates with suspected systemic herpes simplex virus infection.
+
 - Add Ousey 2026 plozasiran ([doi:10.1002/jcph.70190](https://doi.org/10.1002/jcph.70190)) -- adults with familial chylomicronemia syndrome.
 
 - Add Ravva 2010 varenicline exposure-response models ([doi:10.1038/clpt.2009.282](https://doi.org/10.1038/clpt.2009.282)) -- adult cigarette smokers in smoking-cessation trials.